removing crosslinking rounding

+++ bioprint/src/octoprint/server/api/printer.py
@@ -147,20 +147,20 @@ def printerToolCommand():
 			return make_response("Invalid value for flow rate: %s" % str(e), 400)
 
 	elif command == "light":
-		intensity = int(data["intensity"])
+		intensity = data["intensity"]
 		printer.light_intensity(intensity)
 
 	elif command == "crosslink":
 		cl_params = data["cl_params"]
 		if not isinstance(cl_params["cl_layers"], (int)):
 			return make_response("Not a number for cl_layers: %r" % cl_params["cl_layers"], 400)
-		if not isinstance(cl_params["cl_intensity"], (int)):
+		if not isinstance(cl_params["cl_intensity"], (float)):
 			return make_response("Not a number for cl_intensity: %r" % cl_params["cl_intensity"], 400)
 		if not isinstance(cl_params["cl_duration"], (int, long, float)):
 			return make_response("Not a number for cl_duration: %r" % cl_params["cl_duration"], 400)
 		if not isinstance(cl_params["cl_end"], (bool)):
 			return make_response("Not a bool for cl_end: %r" % cl_params["cl_end"], 400)
-		if not isinstance(cl_params["cl_end_intensity"], (int)):
+		if not isinstance(cl_params["cl_end_intensity"], (float)):
 			return make_response("Not a number for cl_end_intensity: %r" % cl_params["cl_end_intensity"], 400)
 		if not isinstance(cl_params["cl_end_duration"], (int, long, float)):
 			return make_response("Not a number for cl_end_duration: %r" % cl_params["cl_end_duration"], 400)
